# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This cashback for buying tariffs",
        "title": "Cashback Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "@xumoyiddin_xolmuminov",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cashback": {
            "post": {
                "description": "Increase cashback for a user by sending a request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Increase cashback",
                "parameters": [
                    {
                        "description": "Cashback request body",
                        "name": "cashback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CashbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    }
                }
            }
        },
        "/cashback/decrease": {
            "post": {
                "description": "Decrease cashback for a user by sending a request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Decrease cashback",
                "parameters": [
                    {
                        "description": "Cashback decrease request body",
                        "name": "decrease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CashbackDecreaseQueue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    }
                }
            }
        },
        "/cashback/{id}": {
            "get": {
                "description": "Get the current cashback information for a Cinerama user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Get cashback by Cinerama user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cinerama User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    }
                }
            }
        },
        "/cashback_history/{id}": {
            "get": {
                "description": "Get full cashback history records for a Cinerama user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cashback"
                ],
                "summary": "Get cashback history by Cinerama user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cinerama User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.CashbackResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requests.CashbackDecreaseQueue": {
            "type": "object",
            "required": [
                "cinerama_user_id",
                "decrease_cashback_amount",
                "device",
                "host_ip"
            ],
            "properties": {
                "cinerama_user_id": {
                    "type": "integer"
                },
                "decrease_cashback_amount": {
                    "type": "number"
                },
                "device": {
                    "type": "string"
                },
                "host_ip": {
                    "type": "string"
                }
            }
        },
        "requests.CashbackRequest": {
            "type": "object",
            "required": [
                "cinerama_user_id",
                "device",
                "host_ip",
                "tariff_price"
            ],
            "properties": {
                "cinerama_user_id": {
                    "type": "integer"
                },
                "device": {
                    "type": "string"
                },
                "host_ip": {
                    "type": "string"
                },
                "tariff_price": {
                    "type": "number"
                }
            }
        },
        "response.CashbackResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}